Move NoteFeatureUtil to the commands.note package.
